-Reworked AvailObjectRepresentation "int[] intSlots" into "long[] longSlots".  Reworked all descriptor classes to use BitFields for all 32-bit access, or intSlot(_,_) and setIntSlot(_,_,_) for variable length int-like fields.  This null refactor is early preparation for a native (non-Java) VM representation that will require 8-byte alignment. -The mechanism for constructing strings and various kinds of tuples was modified to conform with that already implemented in ByteStringTupleDescriptor and TwoByteStringTupleDescriptor, whereby a generator object is provided to supply successive elements.  This allows a grouped-write optimization in classes like NybbleTupleDescriptor, and hides details of representaton from clients.  It's still unclear if this has caused a significant slowdown. -Plenty of other tweaks were made along the way, some related and some not.

+++ src/com/avail/builder/AvailBuilder.java
@@ -2895,7 +2895,8 @@ public final class AvailBuilder
 						position.value += result;
 						if (buffer.hasRemaining())
 						{
-							channel.write(buffer, position.value, null, this);
+							channel.write(
+								buffer, position.value.intValue(), null, this);
 						}
 					}
